Backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
Django==3.2.4
django-cors-headers==3.7.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
idna==2.10
Pillow==8.3.1
pycparser==2.20
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
nsedata==0.1
bsedata==0.5.0
